refactor(cmd-vel): merge looper logic into cmd_vel_control, remove standalone tool
